third_party/blink/web_tests/fast/filesystem/not-enough-arguments-expected.txt

This test tries calling various filesystem functions with too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS webkitResolveLocalFileSystemURL() threw exception TypeError: Failed to execute 'webkitResolveLocalFileSystemURL' on 'Window': 2 arguments required, but only 0 present..
PASS webkitRequestFileSystem() threw exception TypeError: Failed to execute 'webkitRequestFileSystem' on 'Window': 3 arguments required, but only 0 present..
PASS webkitRequestFileSystem(PERSISTENT) threw exception TypeError: Failed to execute 'webkitRequestFileSystem' on 'Window': 3 arguments required, but only 1 present..
PASS webkitRequestFileSystem(PERSISTENT, 100) threw exception TypeError: Failed to execute 'webkitRequestFileSystem' on 'Window': 3 arguments required, but only 2 present..
Successfully obtained Persistent FileSystem:file__0:Temporary
PASS fileSystem.root.removeRecursively() threw exception TypeError: Failed to execute 'removeRecursively' on 'DirectoryEntry': 1 argument required, but only 0 present..
PASS fileSystem.root.moveTo() threw exception TypeError: Failed to execute 'moveTo' on 'Entry': 1 argument required, but only 0 present..
PASS fileSystem.root.copyTo() threw exception TypeError: Failed to execute 'copyTo' on 'Entry': 1 argument required, but only 0 present..
PASS reader.readEntries() threw exception TypeError: Failed to execute 'readEntries' on 'DirectoryReader': 1 argument required, but only 0 present..
PASS entry.createWriter() threw exception TypeError: Failed to execute 'createWriter' on 'FileEntry': 1 argument required, but only 0 present..
PASS entry.file() threw exception TypeError: Failed to execute 'file' on 'FileEntry': 1 argument required, but only 0 present..
PASS writer.write() threw exception TypeError: Failed to execute 'write' on 'FileWriter': 1 argument required, but only 0 present..
PASS writer.seek() threw exception TypeError: Failed to execute 'seek' on 'FileWriter': 1 argument required, but only 0 present..
PASS writer.truncate() threw exception TypeError: Failed to execute 'truncate' on 'FileWriter': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
